Create user controller, update router
