+++ lib/routers/users.js
@@ -1,22 +1,19 @@
 const express = require('express');
-const User = require('../models/User')
 const passport = require('passport')
-const paramValidator = require('../helpers/user/IncomingUserParamValidator')
+const IncomingUserParamValidator = require('../helpers/user/IncomingUserParamValidator')
+const UserController = require('../controllers/UserController')
+const createError = require('http-errors');
 
 const router = express.Router();
 
 router.post('/users', async function(req, res, next) {
   try {
-    new paramValidator(req.body)
+    new IncomingUserParamValidator(req.body)
   } catch (e) {
     res.status(400).json(e)
     next()
   }
-
-  let user = new User(req.body)
-  user.save().then(()=>{
-    res.status(201).send(user.toAuthJSON())
-  }).catch(next)
+  UserController.createUserAction(req.body, res, next)
 })
 
 router.post('/users/login', function(req, res, next) {
@@ -33,33 +30,18 @@ router.post('/users/login', function(req, res, next) {
 })
 
 router.patch('/users/:id', function(req, res, next) {
-  passport.authenticate('jwt', {session: false}, (err, authUser, info) => {
+  passport.authenticate('jwt', {session: false}, (err, user, info) => {
     if(err) {
       return next(err)
     }
-    if(authUser){
-      User.findById(req.params.id, function(err, user){
-        if(err) {
-          return res.status(400).send(err)
-        }
-        if(user.id !== authUser.id) {
-          return res.status(401).json()
-        }
-	try {
-	  verifyProfileInfo(req.body)
-	} catch (err) {
-	  return res.status(400).send(err)
-	}
-        user.setProfile(req.body)
-        user.save()
-        .then(()=>{
-          res.status(204).send()
-        })
-        .catch((err)=>{res.status(500).send(err)})
-      })
+    if(user){
+      if (user.id !== req.params.id ) {
+        next(createError(401, "Not authorized to edit user"))
+      }
+      UserController.updateUserAction(user, req.body, res, next)
       return
     }
-    return res.status(401).json(info)
+    return res.status(403).json(info)
   })(req, res, next)
 })
 
@@ -67,20 +49,4 @@ router.patch('/users/:id', function(req, res, next) {
 //   //get user in regards to user who created membership request.
 // }
 
-const verifyProfileInfo = (profileInfo) => {
-  if (!isValidCoords(profileInfo.location.coords)) {
-    throw 'Bad location coordinates'
-  }
-}
-
-const isValidCoords = ([long, lat]) => {
-  if (long <= 180 &&
-      long >= -180 &&
-      lat <= 90 &&
-      lat >= -90) {
-    return true
-  }
-  
-  return false
-}
 module.exports = router;
